Issue #2794485 by miteshmap, bojanz: Remove EntityStoreInterface, add EntityStoresInterface
